MVC structure in place
